update how the root path is first initialized
Upon startup, PageBuilder checks for a settings.ini file containing the
path to a valid Hypnospace data directory. If that path happens to be
invalid, then any attempts to load a page which references an asset
will fail.

Steam users probably won't run into this problem because PageBuilder will
initialize itself to the Steam folder by default. Anyone using an
alternate distribution such as the GOG version or someone who keeps their
Steam library in a non-standard location will end up with PageBuilder's
current working directory being selected, but this is not a valid
location.

This update does three things:

 - In the case where no settings.ini file already exists, it will check
   whether a "data" folder is present before assuming that the current
   working directory is valid.
 - Make sure that the program's current directory is changed to the data
   folder selected.
 - Add additional error handling for cases where the user selects a folder
   which lacks a "data" subfolder or the program fails to access the
   selected folder.